refactor(theme-doc): share pageGroups analysis
